Fix stats file creation when stats collection off
I noticed that despite having collectStats set to false, upon vault
or plugin load, the .vault-stats file would keep coming back. This is
due to:

In the plugin onload(), a BarManager is created
In the BarManager constructor, a DataManager is created
In the DataManager constructor, the .vault-stats file is created and
then written to, regardless of the user's collectStats setting.

We add an additional check in the DataManager constructor for the
collectStats setting, propagating settings through constructors
as necessary.